PG170906A:	VUEE >= PG170906A
Added a new "standard" type, wint, representing the word-sized (signed)
integer. It is useful for some drivers, as otherwise word-sized signed
integers, e.g., used by accelerometers, cannot be made consistent
across MSP430/ARM.

Modified the join operation, which used to be a simple macro translating
into when (pid, state). That simple way of interpreting join would fail,
say, in (tempting) constructs like "join (runfsm a, state)" if runfsm
failed on malloc returning zero as pid. So now, join is a function
which:

	- returns the pid passed as the first argument, if that argument
	  pointed to a valid thread, and issues when (pid, state)
	- returns 0 otherwise (also in the case when pid was zero to
	  begin with) and does nothing

This way the caller can easily know what has happened.

As an implication, I have removed the clumsy and confusing combination
of "call" macro/function which is now replaced with a series of
statement (generated by picomp) amounting to this:

	if (join (runfsm a (arg), state))
		release;

so the release doesn't happen, if, say, runfsm has failed on malloc. A
similar code is generated for VUEE.

A tentative driver for the mpu9250 accelerometer on CC1350 SENSORTAG.
I still cannot get gyro to work, but the basics, like motion detection
do work. Quite likely, this is all we need for now.

While at it, I redid the handling of I2C for CC1350. The mpu9250 uses
a different SDA/SCL pin pair that the remaining sensors (which share
the same I2C bus) and CC1350 can only handle one bus at a time. While
I could try a raw-pin (software and partial) implementation of I2C for
the mpu9250 pins, I opted instead for a switchover scheme whereby the
I2C bus is reconfigured when communication switches from one to the
other. That makes sense as long as we do not use I2C interrupts (which
we don't).

Things are tricky with the "new generation" of chips. While writing
a driver for, say, bma250, took me one day (and just a few hours of
it), the mpu9250 took a week (and it still doesn't quite work). The
culprit: undocumented features only available through the manufacturer's
driver (messy and utterly incompatible with anything PicOS). This seems
to be the new philosophy for "microcontrollers".