PG170913A:
I have writen a simple driver for the one-bit microphone on the
CC1350 SENSORTAG which seems to work and probably can be used for
things like noise detection.

Fixed a bug in autoconnect.tcl causing a failure to connect to a new
COM port, if the previous set of COM ports was empty.