PG180426A:
I have replaced the old manual with a new version that I wrote
recently (and still keep working on in my spare time). While the
new manual (aka the book) is not complete yet (and hasn't even been
proofread), it looks decent enough to me to do as a replacement for
the old manual, especially that the latter has not been updated for
a while (and, generally, has not been considered very friendly).