PG180601A:
I have changed the signature of RFC_xmt, one of the so-called
assessment methods, to base its operation of the (full) packet,
rather than just the length. This makes more sense, because, in
some advanced models, different components of the same packet can
be sent using different encodings (rates).

A new version of the manual (aka book). Enjoy the index!